+++ src/components/endpoints/SendLyx.vue
@@ -2,7 +2,7 @@
 import Web3Utils from "web3-utils";
 import { getState, setState } from "@/stores";
 import { ref } from "vue";
-import Notifications from "@/components/shared/Notification.vue";
+import Notifications from "@/components/Notification.vue";
 import useNotifications from "@/compositions/useNotifications";
 import { TransactionConfig } from "web3-core";
 import useWeb3 from "@/compositions/useWeb3";
