+++ src/views/Home.vue
@@ -5,7 +5,7 @@ import SetData from "@/components/endpoints/SetData.vue";
 import Permissions from "@/components/endpoints/Permissions.vue";
 import Assets from "@/components/endpoints/Assets.vue";
 import Sign from "@/components/endpoints/Sign.vue";
-import Notifications from "@/components/shared/Notification.vue";
+import Notifications from "@/components/Notification.vue";
 import { computed } from "vue";
 
 const hasExtension = computed(() => {
